#153 refactor skill system

+++ source/battleActions/personas/Io.h
@@ -1,10 +1,10 @@
 #pragma once
 #include "PersonaBase.h"
 
-#include "../skills/AttackSkill.h"
 #include "../skills/Dia.h"
 #include "../skills/Garu.h"
 #include "../skills/Magaru.h"
+#include "../skills/Skill.h"
 
 struct Io : PersonaBase
 {
@@ -29,15 +29,15 @@ struct Io : PersonaBase
         skillCount = 3;
         skills = mySkills;
 
-        battleStats.affinities[Element::Slash] = BattleStats::Neutral;
-        battleStats.affinities[Element::Strike] = BattleStats::Neutral;
-        battleStats.affinities[Element::Pierce] = BattleStats::Neutral;
-        battleStats.affinities[Element::Fire] = BattleStats::Neutral;
-        battleStats.affinities[Element::Ice] = BattleStats::Neutral;
-        battleStats.affinities[Element::Electricity] = BattleStats::Weak;
-        battleStats.affinities[Element::Wind] = BattleStats::Resist;
-        battleStats.affinities[Element::Light] = BattleStats::Neutral;
-        battleStats.affinities[Element::Dark] = BattleStats::Neutral;
+        battleStats.affinities[(u32)Element::Slash] = BattleStats::Neutral;
+        battleStats.affinities[(u32)Element::Strike] = BattleStats::Neutral;
+        battleStats.affinities[(u32)Element::Pierce] = BattleStats::Neutral;
+        battleStats.affinities[(u32)Element::Fire] = BattleStats::Neutral;
+        battleStats.affinities[(u32)Element::Ice] = BattleStats::Neutral;
+        battleStats.affinities[(u32)Element::Electricity] = BattleStats::Weak;
+        battleStats.affinities[(u32)Element::Wind] = BattleStats::Resist;
+        battleStats.affinities[(u32)Element::Light] = BattleStats::Neutral;
+        battleStats.affinities[(u32)Element::Dark] = BattleStats::Neutral;
         // add arcana
     }
 };
